Add the S type variable for “stage.”
This type variable is very important as it will allow us to prevent
expressions, functions, statements, scopes etc. from using other
constructs that are not allowed in their given shader stage. An example
of this is the `gl_VertexID` variable, which is going to be a free
variable that is available only in vertex shaders. Trying to use it in a
fragment shader or anything else should generate a compile error.