Mutating function types and expressions.
We now need to lift the expressions up the EDSL. It’s not going to be
super hard but it will require a bit of manipulation and that’s the
moment where we need to decide whether we want to remove the
`FnBuilder` trait we have in shades or not.

We probably want.